requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
peewee==3.15.4

Flask~=2.2.2
requests~=2.28.2